Add files via upload
This is the working code for using an I2C bus with temperature sensor (Sparkfun TMP102). The code will allow the display of the temperature on attached LCD. If one exceeds the temperature of 80 degrees Fahrenheit, the LCD will also display a 1 below the temperature reading, whereas any temperature under 80 degrees Fahrenheit will display a 0.